+++ src/ui/menu_view.py
@@ -9,6 +9,15 @@ class MenuView:
         self._frame = None
 
         self._initialize()
+    
+    def _handle_button_click(self, n):
+        button_value = n
+        if button_value == 1:
+            print("pushed 1")
+        elif button_value == 2:
+            print("pushed 2")
+        elif button_value == 3:
+            print("pushed 3")
 
     def pack(self):
         self._frame.pack(fill=constants.X)
@@ -16,6 +25,7 @@ class MenuView:
     def destroy(self):
         self._frame.destroy()
     
+
     def _initialize(self):
         self._frame = ttk.Frame(master=self._root)
         label = ttk.Label(master=self._frame, text="this is the main menu")
@@ -40,17 +50,17 @@ class MenuView:
         button_a1 = ttk.Button(
             master=self._frame,
             text="Action1",
-            command=print("action 1 registered")
+            command = lambda: self._handle_button_click(1)
         )
         button_a2 = ttk.Button(
             master=self._frame,
             text="Action2",
-            command=print("action 2 registered")
+            command = lambda: self._handle_button_click(2)
         )
         button_a3 = ttk.Button(
             master=self._frame,
             text="Action3",
-            command=print("action 3 registered")
+            command = lambda: self._handle_button_click(3)
         )
 
         label.grid(row=0, column=0)