+++ dokumentaatio/changelog.md
@@ -3,5 +3,6 @@
 ## Viikko 3
 
 - Käyttöliittymä aloitettu
+- Menut (login, main menu, collection, settings, tarrojenhankintanapit) olemassa
 
 ## Viikko 4